requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
websockets==12.0
python-dotenv==1.0.1
twilio==8.1.0
loguru==0.7.2
groq>=0.5.0
httpx>=0.27.0
